Move File and Block in/out stream tests to integration-tests module and rename them after IntegrationTest
